refactor: getMetaContent
