+++ src/providers/frame/readers/metadata.ts
@@ -1,9 +1,29 @@
 import { compact, last } from 'lodash-es';
 
 import { FrameProtocol } from '@/constants/enum.js';
-import { getFrameMetaContent } from '@/helpers/getFrameMetaContent.js';
 import { q, qsAll } from '@/helpers/q.js';
 import { ActionType, type FrameButton, type FrameInput } from '@/types/frame.js';
+import { getMetaContent } from '@/helpers/getMetaContent.js';
+
+/**
+ * Get content of frame meta tag with frame client protocol
+ * @param document
+ * @param criteria
+ * @returns
+ */
+function getFrameMetaContent(
+    document: Document,
+    criteria: Record<Exclude<ReturnType<typeof getProtocol>, undefined>, string> & { og?: string },
+) {
+    const protocol = getProtocol(document);
+    if (!protocol) return null;
+
+    const selector = criteria[protocol];
+    const content = getMetaContent(document, selector);
+    if (content) return content;
+
+    return criteria.og ? getMetaContent(document, criteria.og) : null;
+}
 
 export function getTitle(document: Document): string | null {
     return (
